Move types related to account to a separate module
No functional change.
