+++ src/types.rs
@@ -1,4 +1,5 @@
 use super::{Indexable, IterableMap, NetworkType, Store, TrafficDirection};
+pub use crate::account::{Account, PasswordHashAlgorithm, Role, SaltedPassword};
 use anyhow::{bail, Context, Result};
 use bincode::Options;
 use chrono::{
@@ -7,10 +8,6 @@ use chrono::{
 use data_encoding::BASE64;
 use flate2::read::GzDecoder;
 use ipnet::IpNet;
-use ring::{
-    digest, pbkdf2,
-    rand::{self, SecureRandom},
-};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 use std::{
@@ -19,11 +16,10 @@ use std::{
     convert::TryFrom,
     io::{BufReader, Read},
     net::IpAddr,
-    num::NonZeroU32,
     ops::RangeInclusive,
     sync::Arc,
 };
-use strum_macros::{Display, EnumString};
+use strum_macros::Display;
 
 pub trait FromKeyValue: Sized {
     fn from_key_value(key: &[u8], value: &[u8]) -> Result<Self>;
@@ -37,27 +33,6 @@ where
         let entry = bincode::DefaultOptions::new().deserialize::<Self>(value)?;
         Ok(entry)
     }
-}
-
-#[derive(Default, Deserialize, Serialize)]
-pub enum PasswordHashAlgorithm {
-    #[default]
-    Pbkdf2HmacSha512 = 0,
-    Argon2id = 1,
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct Account {
-    pub username: String,
-    pub password: SaltedPassword,
-    pub role: Role,
-    pub name: String,
-    pub department: String,
-    pub creation_time: DateTime<Utc>,
-    pub last_signin_time: Option<DateTime<Utc>>,
-    pub allow_access_from: Option<Vec<IpAddr>>,
-    pub max_parallel_sessions: Option<u32>,
-    pub password_hash_algorithm: PasswordHashAlgorithm,
 }
 
 #[derive(
@@ -216,12 +191,6 @@ pub enum DataType {
 pub struct Endpoint {
     pub direction: Option<TrafficDirection>,
     pub network: HostNetworkGroup,
-}
-
-#[derive(Clone, Copy, Deserialize, Serialize)]
-#[repr(u32)]
-pub enum HashAlgorithm {
-    Sha512 = 0,
 }
 
 // `hosts` and `networks` must be kept sorted.
@@ -386,66 +355,6 @@ pub struct Outlier {
     pub event_ids: Vec<i64>,
     pub size: i64,
     pub model_id: i32,
-}
-
-/// Possible role types of `Account`.
-#[derive(Clone, Copy, Debug, Display, Eq, PartialEq, Deserialize, Serialize, EnumString)]
-pub enum Role {
-    #[strum(serialize = "System Administrator")]
-    SystemAdministrator,
-    #[strum(serialize = "Security Administrator")]
-    SecurityAdministrator,
-    #[strum(serialize = "Security Manager")]
-    SecurityManager,
-    #[strum(serialize = "Security Monitor")]
-    SecurityMonitor,
-}
-
-#[derive(Clone, Deserialize, Serialize)]
-pub struct SaltedPassword {
-    salt: Vec<u8>,
-    hash: Vec<u8>,
-    algorithm: HashAlgorithm,
-    iterations: NonZeroU32,
-}
-
-impl SaltedPassword {
-    pub fn new(password: &str) -> Result<Self> {
-        const ITERATIONS: u32 = 100_000;
-
-        let iterations = NonZeroU32::new(ITERATIONS).expect("valid u32");
-        let rng = rand::SystemRandom::new();
-        let mut salt = vec![0_u8; digest::SHA512_OUTPUT_LEN];
-        rng.fill(&mut salt)?;
-        let mut hash = vec![0_u8; digest::SHA512_OUTPUT_LEN];
-        pbkdf2::derive(
-            pbkdf2::PBKDF2_HMAC_SHA512,
-            iterations,
-            &salt,
-            password.as_bytes(),
-            &mut hash,
-        );
-        Ok(Self {
-            salt,
-            hash,
-            algorithm: HashAlgorithm::Sha512,
-            iterations,
-        })
-    }
-
-    #[must_use]
-    pub fn is_match(&self, password: &str) -> bool {
-        match self.algorithm {
-            HashAlgorithm::Sha512 => pbkdf2::verify(
-                pbkdf2::PBKDF2_HMAC_SHA512,
-                self.iterations,
-                &self.salt,
-                password.as_bytes(),
-                &self.hash,
-            )
-            .is_ok(),
-        }
-    }
 }
 
 #[derive(Deserialize)]
